==> cod_code/1_LSB_Lesliematrices_cod_v3.rmd <==
---
title: "Create Leslie & LSB matrices from cod data"
author: "Mikaela Provost"
date: "Feb 17, 2019"
output: html_document
---


```{r}
# ---
# plan:
# 1. read cod data, break in separate populations
# 2. load functions
# 3. execute functions for each population of cod, export Leslie
# 4. execute functions for each population of cod, export LSB matrices 

rm(list=ls())

# ---
# 1. load cod data, break into separate populations
# ---
source("C:/Users/provo/Documents/GitHub/popdy/cod_code/0_load_cod_data.r")
# this loads 'datalist' which has the data for each population
# combined into one list
        
# ---
# 2. load functions
# ---
source("C:/Users/provo/Documents/GitHub/popdy/cod_code/2_cod_functions.r")
  # assemble_Leslie()
  # extract_first_eigen_value()
  # calc_LSB_at_age_by_F()


# ---
# 3. execute functions for each population of cod, export Leslie 
# ---

# prep for looping
tknot = 0
F.halfmax = 0

# loop over pop data in datalist to generate Leslie matrices
Alist = as.list(rep(NA,length(datalist))) # store Leslie matrix
names(Alist) = codNames



# step through each dataset in datalist
for (i in 1:length(names(Alist))) {
  # load parms for cod pop i
  source(file = paste('C:/Users/provo/Documents/GitHub/popdy/cod_pops/',names(Alist)[i], '.r', sep=''))
                       
  # this should load parms: L_inf, K, TEMP, maxage
  out=assemble_Leslie(data=datalist[[i]], maxage=maxage, K=K, L_inf=L_inf, 
                      TEMP=TEMP,F.halfmax=0,tknot=0,B0=B0,B1=B1)
  Alist[[i]]=out$A #output from assemble_Leslie has two matricies, just choose Leslie (A)
  write.table(Alist[[i]],
              file=paste("C:/Users/provo/Documents/GitHub/popdy/cod_code/mikaelaLeslie/matrix_maxages/",
                         codNames[i],".txt",sep=""))
  rm(K,L_inf,maxage,TEMP,B0,B1)
  
}
# the matricies made above are considered 'base' matrix. Have not been adjusted. 


# ---
# 4. execute functions for each population of cod, export LSB matrices 
# ---

#prep for looping
tknot = 0
F.halfmax = 0 #seq(0,3,by=0.2)

# loop over pop data in datalist to generate LSB matrices
LSBlist = as.list(rep(NA,length(codNames))) # store matrix of LSB vs F
names(LSBlist) = codNames
#ks = c(0,0.2,0.50,8,1) # different k values


for (i in 1:length(LSBlist)) { # step through each pop in datalist
   # load parms for cod pop i
  source(file = paste('C:/Users/provo/Documents/GitHub/popdy/cod_pops/',names(LSBlist)[i], '.r', sep=''))
  # this should load parms: L_inf, K, TEMP, maxage
  
  #lsb.at.k = matrix(0,nrow=maxage,ncol=length(ks)) #create empty matrix to store LSB at age at k
  #for (u in 1:length(ks)) {
  lsb.at.k = calculate_LSB_at_age_by_F(maxage=maxage,
                                    L_inf=L_inf, K=K, TEMP=TEMP, F.halfmax=F.halfmax,
                                    B0=B0,B1=B1)
  #}
  LSBlist[[i]] = lsb.at.k
  
  # export table of lsb at age for different k values (need this to calculate new Leslie matricies)
  #write.table(LSBlist[[i]],
  #            file=paste("C:/Users/provo/Documents/GitHub/popdy/cod_code/mikaelaLSB/LSBmatrix_maxages/",
   #                      codNames[i],".txt",sep=""))
   write.table(LSBlist[[i]],
              file=paste("C:/Users/provo/Documents/GitHub/popdy/cod_code/mikaelaLSB/LSBmatrix_maxages_v1/",
                         codNames[i],".txt",sep=""))
  rm(K,L_inf,maxage,TEMP,B0,B1)
}



```
