==> cod_code/13_cross_check_Jac_Leslie_eigenvalues.rmd <==
# Check eigenvalues and dynamics of Jacobian when lambda1 < and > 1
# last updated: 2019 Aug 5



# ===================================================================
# North Sea example
source(file = paste('C:/Users/provo/Documents/GitHub/popdy/cod_pops/','Northsea', '.r', sep=''))
# load functions 
source("C:/Users/provo/Documents/GitHub/popdy/cod_code/2_cod_functions.r")
Leslieout = assemble_Leslie(maxage=maxage, K=K, L_inf=L_inf, TEMP=TEMP,
                            F.halfmax=0, B0=B0, B1=B1, tknot=0)

NEAR = Leslieout$NEAR
LEP = sum(NEAR$egg_production)
1/LEP
A = Leslieout$A
# Convert Leslie to Jacobian: multiply top row of Leslie by 1/(alpha*LEP^2)
# Since 
alpha = 1.5
beta = 1000
initial_eggs = 1000
timesteps = 1000

#before adjusting fecundity
extract_first_eigen_value(A)
#now: adjust fecundity so that LEP=0.95
Aadj = A
conLEP = 0.95
adjFec = round(1/conLEP,digits=1) #1/LEP is 1.1, which means alpha must be greater than this


Aadj[1,] = Aadj[1,]/(LEP*adjFec) #adjusted A
extract_first_eigen_value(Aadj) #what's lambda1 now? 0.98: we made LEP=0.95 which means lambda1<1
A = Aadj #relabel matrix for easiness


#sim_model <- function(A,timesteps,alpha,beta,sig_r,initial_eggs) {
age_at_mat <- min(which(!A[1,] ==0)) # set inequality to min threshold for maturity
maxage = length(A[,1])
ages = length(seq(1,maxage,by=1))
N0 = c(initial_eggs, rep(0,ages-1)) #vector, length=num of ages, first age is initial eggs
set.seed(2) #Set the seed so that every simulation uses same random sequence

Nt = matrix(0,ages,timesteps) #empty matrix to store output: rows=ages, cols=time

# below is code that is slightly modified, old code is above. 
Nt[,1] = N0 #Put in initial values
Nt[,2] = A %*% Nt[,1]  # multiply initial age vector with Leslie to get 2nd age vector
eggs = c(initial_eggs, rep(NA,timesteps-2)) #will save egg production here, this will be the input in BH
eggs[2] = Nt[1,2]
recruits1 = eggs[1]
recruits1 = eggs[1]/( (1/alpha) + (eggs[1]/beta) )
recruits = c(0,rep(NA, timesteps-2)) #will save recruits here (output from BH)
recruits[2] = recruits1 #we are ignoring recorinding recruits at time 0



for(t in 1:(timesteps-2)){ #step through time
  #Save egg production, this is new number of age 1 individuals
  eggs[t+1] = Nt[1,t+1] 
  #save recruits, treat egg production as spawners in BH
  recruits[t+1] = eggs[t]/( (1/alpha) + (eggs[t]/beta) ) #((alpha*eggs[t+1])/(1+beta*eggs[t+1])) 
  #replace age 1 with recruits from BH, add noise    
  Nt[1,t+1] = (recruits[t+1]) *exp(sig_r*rnorm(1,mean=0,sd=1)) 
  #perform population projection for one time step   
  Nt[,t+2] = A %*% Nt[,t+1] 
  
}
#Nsize = colSums(Nt) #total population size, sums num at age for each timestep
Nsize = colSums(Nt[age_at_mat:maxage,]) #sum rows that correspond to spawning adults
plot(Nsize)
N_t = Nt[1,][1:(timesteps-1)] #Nt is number of spawning adults, aka eggs, after environmental noise
eggs = eggs[1:(timesteps-1)] #egg production
recruits = recruits[1:(timesteps-1)] #recruits produced via BH, before influence of environmental noise

return(list(N_t=N_t, eggs=eggs, recruits=recruits, Nsize=Nsize))
}
